# api/rest/friend/openapi.yaml
openapi: 3.1.0
info:
    title: Friend Service API
    description: Friend Service API
    version: 1.0.0
    license:
        name: ISC
        identifier: ISC
servers:
    -   url: 'http://localhost:8080'
security:
    - OAuth2: []
paths:
    /friends/request:
        post:
            operationId: sendFriendRequest
            summary: Send friend request
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                toUserId:
                                    type: string
            responses:
                '200':
                    description: Friend request sent
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/FriendRequest'
                '400':
                    description: Invalid request parameters
                '401':
                    description: Unauthorized
                '403':
                    description: Forbidden
                '500':
                    description: Internal server error
...
components:
    securitySchemes:
        OAuth2:
            type: oauth2
            flows:
                authorizationCode:
                    authorizationUrl: http://localhost:8080/oauth2/authorize
                    tokenUrl: http://localhost:8080/oauth2/token
                    scopes: {}
    schemas:
        FriendRequest:
            type: object
            properties:
                id:
                    type: string
                fromUserId:
                    type: string
                toUserId:
                    type: string
                status:
                    type: string
                    enum: [pending, accepted, rejected]
                createdAt:
                    type: string
                    format: date-string
...